+++ trunk/src/df1b2-separable/df1b2sum.cpp
@@ -52,7 +52,7 @@ void ad_read_pass2_sum(void);
 #endif
    //int nvar=df1b2variable::nvar;
 
-   int total_bytes=3*sizeof(df1b2_header);
+   size_t total_bytes=3*sizeof(df1b2_header);
 // string identifier debug stuff
 #if defined(SAFE_ALL)
   char ids[]="TU";
@@ -184,7 +184,7 @@ void read_pass2_2_sum(void)
   int nvar=df1b2variable::nvar;
   test_smartlist & list=f1b2gradlist->list;
 
-  int total_bytes=3*sizeof(df1b2_header);
+  size_t total_bytes=3*sizeof(df1b2_header);
 // string identifier debug stuff
 #if defined(SAFE_ALL)
   char ids[]="BY";
